src/mem0/requirements.txt
agent-framework==1.0.0
mem0ai
python-dotenv
azure-identity
